Flip the bit on user_not_checked
This changes the variable from true to false when the User event fires
the first time and our watch captures it.

Normally we would not refresh the page again (thereby firing the whole
route aspect again, and possibly refreshing the ng-view), however if the
User is logged in, we DO need to fire it again because we would have
already set the view to the login.html templateUrl and the Login
controller.