Add new Explicit Accept Predicate
This explicit predicate is required because by default if you use the
accept predicate it only checks to see if the accept you want is in the
list sent by the client.

This new predicate will first turn the list of predicates back into
exactly that, a list, AND checks to see it is explicitly in the list.