Move away from having DBSession be threadlocal
Instead add a new request method that grabs a brand new dbsession each
and everytime. This will hopefully make testing simpler as we are no
longer dealing with a global object.

A bunch of clean-up surrounding this still has to be done, as we can now
no longer rely on DBSession existing and having our session.